Reorganized files behind lib and created index to expose the wrapper
