--- config.js
@@ -1,8 +1,0 @@
-const axios = require('axios').default;
-
-const httpService = async request => {
-    const response = await axios(request);
-    return response;
-};
-
-module.exports = httpService;
